requirements.txt
pandas
numpy
sqlalchemy
pymysql
matplotlib
jupyter
python-dotenv